% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scZCL.R
\name{scZCL}
\alias{scZCL}
\title{Zebrafish Cell Landscape mapping
description}
\usage{
scZCL(scdata, numbers_plot = 3)
}
\arguments{
\item{scdata}{data.frame or matrix, col correspond to cells and rows correspond to genes}

\item{numbers_plot}{default is 3, number, it will return top "numbers_plot" records in the plot}
}
\description{
Zebrafish Cell Landscape mapping
description
}
