% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scZCL_gui.R
\name{scZCL_vis}
\alias{scZCL_vis}
\title{Visualize scZCL result in a web browser.}
\usage{
scZCL_vis(zcl_result)
}
\arguments{
\item{zcl_result}{an object generate by scZCL funtion}
}
\description{
Runs interactive \code{shiny} session of \code{scZCL}
}
